Add Studio warning for deprecated course keys
in preparation of dropping support for them entirely.

Re: timing: We will _not_ be going live with this on edx.org at launch;
we'll override this setting on our own installs, initially.

We do, however, want to get this merged ASAP, so that it can still be
pulled into Juniper. That will allow us to drop support in time for the
Koa Named Release, while still providing community operators with a full
Named Release cycle to handle deprecation on their installations.

References:
- [0] TNL-7097